[LCR-53] feat: popup 말줄임 처리
