+++ src/mocks/handlers/noticeModal/NoticeItems.ts
@@ -2,67 +2,68 @@ import { NoticeItemType } from "@/types/NoticeItemType";
 
 export const NoticeItems: NoticeItemType[] = [
   {
-    id: 1,
+    id: 8,
     type: "hot",
-    popup: "BLACKPINK",
-    title: "팝업 포스터",
-    timestamp: "2025-04-26T03:13:00.000Z",
-    stockThreshold: 30,
+    popup: "IVE",
+    title: "아이브 스페셜 카드",
+    timestamp: "2025-04-29T12:00:00.000Z",
+    stockThreshold: 25,
   },
   {
-    id: 2,
+    id: 7,
+    type: "normal",
+    popup: "SEVENTEEN",
+    title: "콘서트 포토북",
+    timestamp: "2025-04-28T10:15:00.000Z",
+    stockThreshold: 40,
+  },
+  {
+    id: 6,
+    type: "hot",
+    popup: "AESPA",
+    title: "스페셜 키링",
+    timestamp: "2025-04-27T08:45:00.000Z",
+    stockThreshold: 20,
+  },
+  {
+    id: 5,
     type: "normal",
     popup: "NEWJEANS",
     title: "한정판 굿즈",
-    timestamp: "2025-04-27T05:30:00.000Z",
+    timestamp: "2025-04-26T05:30:00.000Z",
+    stockThreshold: 10,
+  },
+  {
+    id: 4,
+    type: "normal",
+    popup: "NEWJEANS",
+    title: "한정판 굿즈",
+    timestamp: "2025-04-25T05:30:00.000Z",
     stockThreshold: 10,
   },
   {
     id: 3,
     type: "hot",
-    popup: "AESPA",
-    title: "스페셜 키링",
-    timestamp: "2025-04-28T08:45:00.000Z",
-    stockThreshold: 20,
-  },
-  {
-    id: 4,
-    type: "normal",
-    popup: "SEVENTEEN",
-    title: "콘서트 포토북",
-    timestamp: "2025-04-29T10:15:00.000Z",
-    stockThreshold: 40,
-  },
-  {
-    id: 5,
-    type: "hot",
-    popup: "IVE",
-    title: "아이브 스페셜 카드",
-    timestamp: "2025-04-30T12:00:00.000Z",
-    stockThreshold: 25,
-  },
-  {
-    id: 6,
-    type: "normal",
-    popup: "NEWJEANS",
-    title: "한정판 굿즈",
-    timestamp: "2025-04-27T05:30:00.000Z",
-    stockThreshold: 10,
-  },
-  {
-    id: 7,
-    type: "hot",
-    popup: "BLACKPINK special YEAHYEAH",
-    title: "팝업 포스터",
-    timestamp: "2025-04-26T03:13:00.000Z",
+    popup: "BLACKPINK",
+    title:
+      "팝업 포스터 팝업 포스터 팝업 포스터 팝업 포스터 팝업 포스터 팝업 포스터",
+    timestamp: "2025-04-24T03:13:00.000Z",
     stockThreshold: 30,
   },
   {
-    id: 8,
+    id: 2,
+    type: "hot",
+    popup: "BLACKPINK special YEAHYEAH",
+    title: "팝업 포스터",
+    timestamp: "2025-04-23T03:13:00.000Z",
+    stockThreshold: 30,
+  },
+  {
+    id: 1,
     type: "normal",
     popup: "NEWJEANS",
     title: "한정판 굿즈",
-    timestamp: "2025-04-27T05:30:00.000Z",
+    timestamp: "2025-04-22T05:30:00.000Z", // (이건 시간은 같지만 id가 뒤니까 마지막)
     stockThreshold: 10,
   },
 ];
